Move editor to editor folder
